backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
python-docx
openai
pydantic-settings
reportlab
